Eigen::Map doesn't allow .resize()+disable inplace
Eigen FFT uses resize sometimes so Eigen::Matrices need to be used.
dynamically allocating is more elegant anyway than trying to use maps.

This change also revealed a fundamental issue for inplace
transformations: if HalfSpectrum is enabled, Eigen will try to resize
the dst matrix. If that happens to be the same as src as is the case in
inplace transformations, it will also resize the source, causing
unexpected behaviour. It seems in general that inplace transformations
are simply not intended with the current API; there's no good way to do
it except to not use eigen objects at all as params.